init: add typescript $ necessary deps
