+++ package.json
@@ -1,21 +1,24 @@
 {
-    "name": "ws_task",
-    "version": "1.0.0",
-    "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
-    "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "Andrei Auchynnikau",
-    "keywords": [],
-    "license": "ISC",
-    "devDependencies": {
-        "nodemon": "^2.0.16"
-    },
-    "dependencies": {
-        "ws": "^8.8.0"
-    }
+  "name": "ws_task",
+  "version": "1.0.0",
+  "description": "wss learning",
+  "main": "index.js",
+  "type": "commonjs",
+  "scripts": {
+    "start": "node ./index.js",
+    "start:dev": "ts-node-dev --respawn index.ts",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "SAK74",
+  "keywords": [],
+  "license": "ISC",
+  "devDependencies": {
+    "@types/node": "^20.4.0",
+    "@types/ws": "^8.5.5",
+    "ts-node-dev": "^2.0.0"
+  },
+  "dependencies": {
+    "typescript": "^5.1.6",
+    "ws": "^8.8.0"
+  }
 }
